requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.0.6
django-admin-searchable-dropdown==1.2
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==24.2
django-import-export==2.8.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
idna==3.7
odfpy==1.4.1
openpyxl==3.1.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.5.0
tablib==3.6.1
typing_extensions==4.11.0
urllib3==2.2.1
xlrd==2.0.1
xlwt==1.3.0
